pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shelfard"
version = "0.1.0"
description = "Schema drift detection for REST APIs and databases"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "requests>=2.28",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "mcp>=1.7",
    "langchain-mcp-adapters>=0.1",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9"]

[project.urls]
Homepage = "https://github.com/Ilya225/shelfard"
Repository = "https://github.com/Ilya225/shelfard"

[project.scripts]
shelfard = "shelfard.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["shelfard*"]
